Remove unnecessary routes
As a backend api, the only times there is a need for get requests to
the postController is to 'posts' for all posts in the database,
and 'post_detail' for a single post. Otherwise, only post requests need
to be specified here. Getting forms and such is a frontend job.