Remove/rename modules
The new names show more consistency throughout the app. Also, the
previous overabundance of modules was confusing and hard to follow.
By using the same modules, but passing in a boolean to indicate whether
a published post or draft is needed, repition is reduced and things are
easier to follow overall.