Rename new-draft.html and add page headers
Draft-form allows us to logically use the same page when editing a draft
without needing to create a seperate page for editing and creating.
One reason this may not work is if it isn't possible to pass along info
to the draft-page.html page so that it's fields can be filled in with
the content to be edited. If that is the case, we may need to simply
render the drafts.html page with a form and the filled out fields.
//
Also, it was difficult to determine before where you were at a given
time, so headers were added to give a better user experience.